Update to move to more click cli components
